Tutorial Panel added to main menu and two game scenes
Question mark on the left top corner is the help bottom used to open the tutorial page, and this button shown in main menu and both game scenes. When the tutorial page is opening, the time scale become zero, so the game is paused during the time when the tutorial is on.

First try:  Click the bottom to open the tutorial page and close the page with click again.

** Now what we have: H key is used to open and close the tutorial now! I found that we could use both arrow   keys and mouse to choose the items so I change to use the key otherwise the mouse click will mess up with spacebar.

**made an simple image for the background. Replaced the original one, and might make a cover (poster?) if time enough.